Add telemetry to API (#451)
Adds telemetry event logging to `mlem.api` commands. Some Qs:

1) Should we log API event if it was used by CLI command (in this PR we
do not)
2) What if one API command calls some nested API command? Should both
events be logged, or only topmost? (kinda the same as 1) if you really
think about it)
3) What parameters we want to log from each command (will post
suggestions as a separate comment)
4) Should `save, load, load_meta` be added too? It can show interesting
stuff like what types of models and datasets are used more frequently
5) Need a framework to enforce same extra fields for similar API/CLI
calls, but that depends on previous Qs