Read MLEM metafiles with `description` and `labels` (#630)
We deprecated them a while but were able to read them still, but not so
long ago a change in MLEM made these files unreadable. Some repos still
have those, such as
https://github.com/iterative/demo-bank-customer-churn, which is needed
to be read in Studio now.

Besides, I'm switching `try_migrations=True` to automatically try to
read and fix old metafile formats. @mike0sv, you initially set it to
`False` - did you have any specific concerns in mind for that?

cc @amritghimire @jellebouwman - once this is merged and released we
need to check if the MLEM models in the repo can be read without any
troubles. I'll ask your help with that @amritghimire